+++ api/model/api.go
@@ -519,7 +519,6 @@ type PinnedPublicKeySecret struct {
 
 // APIDefinitionSpec represents the configuration for a single proxied API and it's versions.
 type APIDefinitionSpec struct {
-
 	// For server use only, do not use
 	ID string `json:"id,omitempty"`
 
@@ -581,7 +580,6 @@ type APIDefinitionSpec struct {
 	// BasicAuth                  BasicAuthMeta         `json:"basic_auth"`
 	// UseMutualTLSAuth           bool                  `json:"use_mutual_tls_auth"`
 	// ClientCertificates         []string              `json:"client_certificates"`
-	// UpstreamCertificates       map[string]string     `json:"upstream_certificates"`
 
 	// PinnedPublicKeys allows you to whitelist public keys used to generate certificates, so you will be protected in
 	// case an upstream certificate is compromised. Please use PinnedPublicKeysSecretNames if using cert-manager.
@@ -590,6 +588,16 @@ type APIDefinitionSpec struct {
 	// PinnedPublicKeysSecretNames represents the names of the secrets that the controller should look for in the current
 	// namespace which contain the certificates for Certificate Pinning feature.
 	PinnedPublicKeysSecretNames map[string]PinnedPublicKeySecret `json:"pinned_public_keys_secret_names,omitempty"`
+
+	// UpstreamCertificates is a map of domains and certificate IDs that is used by the Tyk
+	// Gateway to provide mTLS support for upstreams
+	UpstreamCertificates map[string]string `json:"upstream_certificates,omitempty"`
+
+	// UpstreamCertificateRefs is a map of domains and secret names that is used internally
+	// to obtain certificates from secrets in order to establish mTLS support for upstreams
+	UpstreamCertificateRefs map[string]string `json:"upstream_certificate_refs,omitempty"`
+
+	// PinnedPublicKeys           map[string]string     `json:"pinned_public_keys"`
 
 	// EnableJWT set JWT as the access method for this API.
 	EnableJWT bool `json:"enable_jwt,omitempty"`
@@ -853,7 +861,6 @@ type ProxyTransport struct {
 // CORS cors settings
 
 type CORS struct {
-
 	// Enable when set to true it enables the cors middleware for the api
 	Enable bool `json:"enable,omitempty"`
 
@@ -925,7 +932,6 @@ type BasicAuthMeta struct {
 }
 
 type OAuth2Meta struct {
-
 	// AllowedAccessTypes are an array of allowable access types.
 	AllowedAccessTypes []AccessTypeEnum `json:"allowed_access_types"` // osin.AccessRequestType